Correct provider site information displayed

+++ src/SFA.Apprenticeships.Infrastructure.TacticalDataServices/Models/ProviderSite.cs
@@ -17,8 +17,8 @@
         public string PostCode { get; set; }
         public int LocalAuthorityId { get; set; }
         public int ManagingAreaID { get; set; }
-        public decimal Longitude { get; set; }
-        public decimal Latitude { get; set; }
+        public double Longitude { get; set; }
+        public double Latitude { get; set; }
         public int GeocodeEasting { get; set; }
         public int GeocodeNorthing { get; set; }
         public string OwnerOrganisation { get; set; }
